flask_server/requirements.txt
psycopg2-binary==2.9.9
Flask==3.0.1